pythonApp/requirements.txt
requests
cython
pyjnius
flask